remove login page; login status now shown on main page
