+++ .run/HH backend.run.xml
@@ -7,9 +7,10 @@
     <envs>
       <env name="PYTHONUNBUFFERED" value="1" />
       <env name="BYPASS_POETRY_SHELL" value="1" />
+      <env name="FLASK_ENV" value="development" />
     </envs>
     <option name="SDK_HOME" value="" />
-    <option name="SDK_NAME" value="Python 3.9 (hardback-hax0r)" />
+    <option name="SDK_NAME" value="Python 3.9 virtualenv at ~/projects/hardback-hax0r/hh-env" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/" />
     <option name="IS_MODULE_SDK" value="false" />
     <option name="ADD_CONTENT_ROOTS" value="true" />
