+++ frontend/src/App.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import MainPage from "./MainPage"; // We'll create this from your existing App component
 import EntryPage from "./EntryPage";
-import LoginPage from "./LoginPage";
 import config from "./config";
 import { AuthProvider } from "./AuthContext";
 
@@ -17,7 +16,6 @@ const App: React.FC = () => {
         <Routes>
           <Route path="/" element={<EntryPage />} />
           <Route path="/main" element={<MainPage />} />
-          <Route path="/login" element={<LoginPage />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
       </BrowserRouter>
